:construction: WiP of the quiz taking screen
After dabbling some this evening, it's getting a bit late so I want to save
this as a work in progress. This is the start of the quiz-taking screen.
Here I'm toying with the idea of having multiple containers that get swapped
in and out of visibility, first with loading then second with showing the
questions.

Right now I just dump the questions in a Pretty so I can see that they're
loaded. Another evening I'll get them into an actual "let's ask some stuff"
state.