:sparkles: Start to flesh out how I'll actually show answers
Using RadioSet was fine for a quick test of things; but it's not a great
experience in terms of UI and doesn't fit in with the feel of the app
either. So here I'm starting to build up both an Answer widget, which shows
a single answer, and an Answers widget, that will implement a method of
showing all the answers and letting the user select one.

This will be keyboard-first; but I think I'd like to add mouse support too.